Method/more_uav/requirements.txt
accelerate>=0.30.0
nltk>=3.8.1
peft>=0.11.0
Pillow>=10.0.0
safetensors>=0.4.0
torch>=2.1.0
tqdm>=4.66.0
transformers>=4.56.2
